Add fallback when no matching level stem is found
Store the world key in level data in order
to use it as fallback to retrieve the level
stem. This can happen when worlds are moved
or copied and the new world key would not be
present in the old level data.

Addionally, when using the Sponge WorldManager
methods to move or copy a world, automatically
update the relevant entries.