Less invasive IO worker queue clear
After ccfca26 we now clear the whole IO worker
queue on world deletion to help with latency on worst
case scenarios. Unfortunately combining this with
0e7fe7e causes race condition where we end up
clearing the synchronize task too.

To fix this, changing the aproach to be less invasive
and instead just halt all storage tasks but still permit
other operations to be performed normally.

By just halting the storage operations we still significantly
reduce the latency. As the world unloading is now async,
the extra in-flight loads don't cause any big concerns.

We also now cleanly clean up the promises in case
anyone is depending on them.