Fix logBlockDrops, attempt two
Commit 3be5556 fixed a state corruption when
the transaction was absorded and would incorrectly
attach a new side effect to the wrong transaction.

Usually this is a sign of a bug but in this case
we were actually taking a dependency on this hidden
quirk where we usually attached it to the correct
transaction.

Rework this by finding the parent transaction and
attach the side-effects directly to it, not to the
tail transaction.